build: update dependencies (pages gem 212 >>> 214) (#172)
I've got a nice dependabot alert about an outdated version of Kramdown. Figured out that in order to update that we'll have to update the github pages gem.

In addition, I made a few small tweaks to the jekyll config such that it complains less when building/serving the page.